+++ test/integration/undo.hh
@@ -48,6 +48,15 @@ class UndoTest : public QTestFixture
   /// \brief Test undoing align commands.
   private slots: void UndoAlign();
 
+  /// \brief Test undoing reset time commands.
+  private slots: void UndoResetTime();
+
+  /// \brief Test undoing reset world commands.
+  private slots: void UndoResetWorld();
+
+  /// \brief Test undoing reset model poses commands.
+  private slots: void UndoResetModelPoses();
+
   /// \brief UndoRedo message received
   /// \param[in] _msg Message containing an undo/redo request.
   private: void OnUndoRedo(ConstUndoRedoPtr &_msg);